+++ autogpt_platform/backend/backend/blocks/test/test_smart_decision_maker.py
