+++ autogpt_platform/backend/backend/util/service_test.py
