--- autogpt_platform/backend/test/__init__.py
@@ -1,3 +1,0 @@
-import os
-
-os.environ["ENABLE_AUTH"] = "false"
