+++ autogpt_platform/frontend/src/components/integrations/credentials-input.tsx
@@ -1,12 +1,8 @@
 import { FC, useEffect, useMemo, useState } from "react";
-import { z } from "zod";
 import { cn } from "@/lib/utils";
-import { useForm } from "react-hook-form";
-import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
 import SchemaTooltip from "@/components/SchemaTooltip";
 import useCredentials from "@/hooks/useCredentials";
-import { zodResolver } from "@hookform/resolvers/zod";
 import { NotionLogoIcon } from "@radix-ui/react-icons";
 import {
   FaDiscord,
@@ -24,22 +20,6 @@ import {
 } from "@/lib/autogpt-server-api/types";
 import { IconKey, IconKeyPlus, IconUserPlus } from "@/components/ui/icons";
 import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog";
-import {
-  Form,
-  FormControl,
-  FormDescription,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import {
   Select,
   SelectContent,
   SelectItem,
@@ -48,6 +28,11 @@ import {
   SelectValue,
 } from "@/components/ui/select";
 import { useBackendAPI } from "@/lib/autogpt-server-api/context";
+import { getHostFromUrl } from "@/lib/utils/url";
+import { APIKeyCredentialsModal } from "./api-key-credentials-modal";
+import { UserPasswordCredentialsModal } from "./user-password-credentials-modal";
+import { HostScopedCredentialsModal } from "./host-scoped-credentials-modal";
+import { OAuth2FlowWaitingModal } from "./oauth2-flow-waiting-modal";
 
 const fallbackIcon = FaKey;
 
@@ -150,16 +135,6 @@ export const CredentialsInput: FC<{
       onSelectCredentials(undefined);
     }
   }, [credentials, selectedCredentials, onSelectCredentials]);
-
-  // Helper function to extract host from URL
-  const getHostFromUrl = (url: string): string | null => {
-    try {
-      const urlObj = new URL(url);
-      return urlObj.hostname;
-    } catch {
-      return null;
-    }
-  };
 
   // Get current host from siblingInputs
   const currentUrl = siblingInputs?.url;
@@ -533,510 +508,5 @@ export const CredentialsInput: FC<{
         <div className="mt-2 text-red-500">Error: {oAuthError}</div>
       )}
     </div>
-  );
-};
-
-export const APIKeyCredentialsModal: FC<{
-  schema: BlockIOCredentialsSubSchema;
-  open: boolean;
-  onClose: () => void;
-  onCredentialsCreate: (creds: CredentialsMetaInput) => void;
-  siblingInputs?: Record<string, any>;
-}> = ({ schema, open, onClose, onCredentialsCreate, siblingInputs }) => {
-  const credentials = useCredentials(schema, siblingInputs);
-
-  const formSchema = z.object({
-    apiKey: z.string().min(1, "API Key is required"),
-    title: z.string().min(1, "Name is required"),
-    expiresAt: z.string().optional(),
-  });
-
-  const form = useForm<z.infer<typeof formSchema>>({
-    resolver: zodResolver(formSchema),
-    defaultValues: {
-      apiKey: "",
-      title: "",
-      expiresAt: "",
-    },
-  });
-
-  if (!credentials || credentials.isLoading || !credentials.supportsApiKey) {
-    return null;
-  }
-
-  const { provider, providerName, createAPIKeyCredentials } = credentials;
-
-  async function onSubmit(values: z.infer<typeof formSchema>) {
-    const expiresAt = values.expiresAt
-      ? new Date(values.expiresAt).getTime() / 1000
-      : undefined;
-    const newCredentials = await createAPIKeyCredentials({
-      api_key: values.apiKey,
-      title: values.title,
-      expires_at: expiresAt,
-    });
-    onCredentialsCreate({
-      provider,
-      id: newCredentials.id,
-      type: "api_key",
-      title: newCredentials.title,
-    });
-  }
-
-  return (
-    <Dialog
-      open={open}
-      onOpenChange={(open) => {
-        if (!open) onClose();
-      }}
-    >
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>Add new API key for {providerName}</DialogTitle>
-          {schema.description && (
-            <DialogDescription>{schema.description}</DialogDescription>
-          )}
-        </DialogHeader>
-
-        <Form {...form}>
-          <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
-            <FormField
-              control={form.control}
-              name="apiKey"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>API Key</FormLabel>
-                  {schema.credentials_scopes && (
-                    <FormDescription>
-                      Required scope(s) for this block:{" "}
-                      {schema.credentials_scopes?.map((s, i, a) => (
-                        <span key={i}>
-                          <code>{s}</code>
-                          {i < a.length - 1 && ", "}
-                        </span>
-                      ))}
-                    </FormDescription>
-                  )}
-                  <FormControl>
-                    <Input
-                      type="password"
-                      placeholder="Enter API key..."
-                      {...field}
-                    />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            <FormField
-              control={form.control}
-              name="title"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Name</FormLabel>
-                  <FormControl>
-                    <Input
-                      type="text"
-                      placeholder="Enter a name for this API key..."
-                      {...field}
-                    />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            <FormField
-              control={form.control}
-              name="expiresAt"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Expiration Date (Optional)</FormLabel>
-                  <FormControl>
-                    <Input
-                      type="datetime-local"
-                      placeholder="Select expiration date..."
-                      {...field}
-                    />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            <Button type="submit" className="w-full">
-              Save & use this API key
-            </Button>
-          </form>
-        </Form>
-      </DialogContent>
-    </Dialog>
-  );
-};
-
-export const UserPasswordCredentialsModal: FC<{
-  schema: BlockIOCredentialsSubSchema;
-  open: boolean;
-  onClose: () => void;
-  onCredentialsCreate: (creds: CredentialsMetaInput) => void;
-  siblingInputs?: Record<string, any>;
-}> = ({ schema, open, onClose, onCredentialsCreate, siblingInputs }) => {
-  const credentials = useCredentials(schema, siblingInputs);
-
-  const formSchema = z.object({
-    username: z.string().min(1, "Username is required"),
-    password: z.string().min(1, "Password is required"),
-    title: z.string().min(1, "Name is required"),
-  });
-
-  const form = useForm<z.infer<typeof formSchema>>({
-    resolver: zodResolver(formSchema),
-    defaultValues: {
-      username: "",
-      password: "",
-      title: "",
-    },
-  });
-
-  if (
-    !credentials ||
-    credentials.isLoading ||
-    !credentials.supportsUserPassword
-  ) {
-    return null;
-  }
-
-  const { provider, providerName, createUserPasswordCredentials } = credentials;
-
-  async function onSubmit(values: z.infer<typeof formSchema>) {
-    const newCredentials = await createUserPasswordCredentials({
-      username: values.username,
-      password: values.password,
-      title: values.title,
-    });
-    onCredentialsCreate({
-      provider,
-      id: newCredentials.id,
-      type: "user_password",
-      title: newCredentials.title,
-    });
-  }
-
-  return (
-    <Dialog
-      open={open}
-      onOpenChange={(open) => {
-        if (!open) onClose();
-      }}
-    >
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>
-            Add new username & password for {providerName}
-          </DialogTitle>
-        </DialogHeader>
-        <Form {...form}>
-          <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
-            <FormField
-              control={form.control}
-              name="username"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Username</FormLabel>
-                  <FormControl>
-                    <Input
-                      type="text"
-                      placeholder="Enter username..."
-                      {...field}
-                    />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            <FormField
-              control={form.control}
-              name="password"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Password</FormLabel>
-                  <FormControl>
-                    <Input
-                      type="password"
-                      placeholder="Enter password..."
-                      {...field}
-                    />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            <FormField
-              control={form.control}
-              name="title"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Name</FormLabel>
-                  <FormControl>
-                    <Input
-                      type="text"
-                      placeholder="Enter a name for this user login..."
-                      {...field}
-                    />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            <Button type="submit" className="w-full">
-              Save & use this user login
-            </Button>
-          </form>
-        </Form>
-      </DialogContent>
-    </Dialog>
-  );
-};
-
-export const HostScopedCredentialsModal: FC<{
-  schema: BlockIOCredentialsSubSchema;
-  open: boolean;
-  onClose: () => void;
-  onCredentialsCreate: (creds: CredentialsMetaInput) => void;
-  siblingInputs?: Record<string, any>;
-}> = ({ schema, open, onClose, onCredentialsCreate, siblingInputs }) => {
-  const credentials = useCredentials(schema, siblingInputs);
-
-  // Helper function to extract host from URL
-  const getHostFromUrl = (url: string): string | null => {
-    try {
-      const urlObj = new URL(url);
-      return urlObj.hostname;
-    } catch {
-      return null;
-    }
-  };
-
-  // Get current host from siblingInputs
-  const currentUrl = siblingInputs?.url;
-  const currentHost = currentUrl ? getHostFromUrl(currentUrl) : "";
-
-  const formSchema = z.object({
-    host: z.string().min(1, "Host is required"),
-    title: z.string().optional(),
-    headers: z.record(z.string()).optional(),
-  });
-
-  const form = useForm<z.infer<typeof formSchema>>({
-    resolver: zodResolver(formSchema),
-    defaultValues: {
-      host: currentHost || "",
-      title: currentHost || "Manual Entry",
-      headers: {},
-    },
-  });
-
-  const [headerPairs, setHeaderPairs] = useState<
-    Array<{ key: string; value: string }>
-  >([{ key: "", value: "" }]);
-
-  // Update form values when siblingInputs change
-  useEffect(() => {
-    if (currentHost) {
-      form.setValue("host", currentHost);
-      form.setValue("title", currentHost);
-    } else {
-      // Reset to empty when no current host
-      form.setValue("host", "");
-      form.setValue("title", "Manual Entry");
-    }
-  }, [currentHost, form]);
-
-  if (
-    !credentials ||
-    credentials.isLoading ||
-    !credentials.supportsHostScoped
-  ) {
-    return null;
-  }
-
-  const { provider, providerName, createHostScopedCredentials } = credentials;
-
-  const addHeaderPair = () => {
-    setHeaderPairs([...headerPairs, { key: "", value: "" }]);
-  };
-
-  const removeHeaderPair = (index: number) => {
-    if (headerPairs.length > 1) {
-      setHeaderPairs(headerPairs.filter((_, i) => i !== index));
-    }
-  };
-
-  const updateHeaderPair = (
-    index: number,
-    field: "key" | "value",
-    value: string,
-  ) => {
-    const newPairs = [...headerPairs];
-    newPairs[index][field] = value;
-    setHeaderPairs(newPairs);
-  };
-
-  async function onSubmit(values: z.infer<typeof formSchema>) {
-    // Convert header pairs to object, filtering out empty pairs
-    const headers = headerPairs.reduce(
-      (acc, pair) => {
-        if (pair.key.trim() && pair.value.trim()) {
-          acc[pair.key.trim()] = pair.value.trim();
-        }
-        return acc;
-      },
-      {} as Record<string, string>,
-    );
-
-    const newCredentials = await createHostScopedCredentials({
-      host: values.host,
-      title: currentHost || values.host,
-      headers,
-    });
-
-    onCredentialsCreate({
-      provider,
-      id: newCredentials.id,
-      type: "host_scoped",
-      title: newCredentials.title,
-    });
-  }
-
-  return (
-    <Dialog
-      open={open}
-      onOpenChange={(open) => {
-        if (!open) onClose();
-      }}
-    >
-      <DialogContent className="max-h-[90vh] max-w-2xl overflow-y-auto">
-        <DialogHeader>
-          <DialogTitle>Add sensitive headers for {providerName}</DialogTitle>
-          {schema.description && (
-            <DialogDescription>{schema.description}</DialogDescription>
-          )}
-        </DialogHeader>
-
-        <Form {...form}>
-          <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
-            <FormField
-              control={form.control}
-              name="host"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Host Pattern</FormLabel>
-                  <FormDescription>
-                    {currentHost
-                      ? "Auto-populated from the URL field. Headers will be applied to requests to this host."
-                      : "Enter the host/domain to match against request URLs (e.g., api.example.com)."}
-                  </FormDescription>
-                  <FormControl>
-                    <Input
-                      type="text"
-                      readOnly={!!currentHost}
-                      placeholder={
-                        currentHost
-                          ? undefined
-                          : "Enter host (e.g., api.example.com)"
-                      }
-                      {...field}
-                    />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-
-            <div className="space-y-2">
-              <FormLabel>Headers</FormLabel>
-              <FormDescription>
-                Add sensitive headers (like Authorization, X-API-Key) that
-                should be automatically included in requests to the specified
-                host.
-              </FormDescription>
-
-              {headerPairs.map((pair, index) => (
-                <div key={index} className="flex items-end gap-2">
-                  <div className="flex-1">
-                    <Input
-                      placeholder="Header name (e.g., Authorization)"
-                      value={pair.key}
-                      onChange={(e) =>
-                        updateHeaderPair(index, "key", e.target.value)
-                      }
-                    />
-                  </div>
-                  <div className="flex-1">
-                    <Input
-                      type="password"
-                      placeholder="Header value (e.g., Bearer token123)"
-                      value={pair.value}
-                      onChange={(e) =>
-                        updateHeaderPair(index, "value", e.target.value)
-                      }
-                    />
-                  </div>
-                  <Button
-                    type="button"
-                    variant="outline"
-                    size="sm"
-                    onClick={() => removeHeaderPair(index)}
-                    disabled={headerPairs.length === 1}
-                  >
-                    Remove
-                  </Button>
-                </div>
-              ))}
-
-              <Button
-                type="button"
-                variant="outline"
-                size="sm"
-                onClick={addHeaderPair}
-                className="w-full"
-              >
-                Add Another Header
-              </Button>
-            </div>
-
-            <Button type="submit" className="w-full">
-              Save & use these credentials
-            </Button>
-          </form>
-        </Form>
-      </DialogContent>
-    </Dialog>
-  );
-};
-
-export const OAuth2FlowWaitingModal: FC<{
-  open: boolean;
-  onClose: () => void;
-  providerName: string;
-}> = ({ open, onClose, providerName }) => {
-  return (
-    <Dialog
-      open={open}
-      onOpenChange={(open) => {
-        if (!open) onClose();
-      }}
-    >
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>
-            Waiting on {providerName} sign-in process...
-          </DialogTitle>
-          <DialogDescription>
-            Complete the sign-in process in the pop-up window.
-            <br />
-            Closing this dialog will cancel the sign-in process.
-          </DialogDescription>
-        </DialogHeader>
-      </DialogContent>
-    </Dialog>
   );
 };
