+++ autogpt_platform/frontend/src/components/integrations/credentials-provider.tsx
@@ -11,6 +11,7 @@ import {
   UserPasswordCredentials,
 } from "@/lib/autogpt-server-api";
 import { useBackendAPI } from "@/lib/autogpt-server-api/context";
+import { useToastOnFail } from "@/components/ui/use-toast";
 
 // Get keys from CredentialsProviderName type
 const CREDENTIALS_PROVIDER_NAMES = Object.values(
@@ -116,6 +117,7 @@ export default function CredentialsProvider({
     useState<CredentialsProvidersContextType | null>(null);
   const { isLoggedIn } = useSupabase();
   const api = useBackendAPI();
+  const onFailToast = useToastOnFail();
 
   const addCredentials = useCallback(
     (
@@ -144,11 +146,16 @@ export default function CredentialsProvider({
       code: string,
       state_token: string,
     ): Promise<CredentialsMetaResponse> => {
-      const credsMeta = await api.oAuthCallback(provider, code, state_token);
-      addCredentials(provider, credsMeta);
-      return credsMeta;
+      try {
+        const credsMeta = await api.oAuthCallback(provider, code, state_token);
+        addCredentials(provider, credsMeta);
+        return credsMeta;
+      } catch (error) {
+        onFailToast("complete OAuth authentication")(error);
+        throw error;
+      }
     },
-    [api, addCredentials],
+    [api, addCredentials, onFailToast],
   );
 
   /** Wraps `BackendAPI.createAPIKeyCredentials`, and adds the result to the internal credentials store. */
@@ -157,14 +164,19 @@ export default function CredentialsProvider({
       provider: CredentialsProviderName,
       credentials: APIKeyCredentialsCreatable,
     ): Promise<CredentialsMetaResponse> => {
-      const credsMeta = await api.createAPIKeyCredentials({
-        provider,
-        ...credentials,
-      });
-      addCredentials(provider, credsMeta);
-      return credsMeta;
+      try {
+        const credsMeta = await api.createAPIKeyCredentials({
+          provider,
+          ...credentials,
+        });
+        addCredentials(provider, credsMeta);
+        return credsMeta;
+      } catch (error) {
+        onFailToast("create API key credentials")(error);
+        throw error;
+      }
     },
-    [api, addCredentials],
+    [api, addCredentials, onFailToast],
   );
 
   /** Wraps `BackendAPI.createUserPasswordCredentials`, and adds the result to the internal credentials store. */
@@ -173,14 +185,19 @@ export default function CredentialsProvider({
       provider: CredentialsProviderName,
       credentials: UserPasswordCredentialsCreatable,
     ): Promise<CredentialsMetaResponse> => {
-      const credsMeta = await api.createUserPasswordCredentials({
-        provider,
-        ...credentials,
-      });
-      addCredentials(provider, credsMeta);
-      return credsMeta;
+      try {
+        const credsMeta = await api.createUserPasswordCredentials({
+          provider,
+          ...credentials,
+        });
+        addCredentials(provider, credsMeta);
+        return credsMeta;
+      } catch (error) {
+        onFailToast("create user/password credentials")(error);
+        throw error;
+      }
     },
-    [api, addCredentials],
+    [api, addCredentials, onFailToast],
   );
 
   /** Wraps `BackendAPI.createHostScopedCredentials`, and adds the result to the internal credentials store. */
@@ -189,14 +206,19 @@ export default function CredentialsProvider({
       provider: CredentialsProviderName,
       credentials: HostScopedCredentialsCreatable,
     ): Promise<CredentialsMetaResponse> => {
-      const credsMeta = await api.createHostScopedCredentials({
-        provider,
-        ...credentials,
-      });
-      addCredentials(provider, credsMeta);
-      return credsMeta;
+      try {
+        const credsMeta = await api.createHostScopedCredentials({
+          provider,
+          ...credentials,
+        });
+        addCredentials(provider, credsMeta);
+        return credsMeta;
+      } catch (error) {
+        onFailToast("create host-scoped credentials")(error);
+        throw error;
+      }
     },
-    [api, addCredentials],
+    [api, addCredentials, onFailToast],
   );
 
   /** Wraps `BackendAPI.deleteCredentials`, and removes the credentials from the internal store. */
@@ -208,26 +230,31 @@ export default function CredentialsProvider({
     ): Promise<
       CredentialsDeleteResponse | CredentialsDeleteNeedConfirmationResponse
     > => {
-      const result = await api.deleteCredentials(provider, id, force);
-      if (!result.deleted) {
-        return result;
-      }
-      setProviders((prev) => {
-        if (!prev || !prev[provider]) return prev;
+      try {
+        const result = await api.deleteCredentials(provider, id, force);
+        if (!result.deleted) {
+          return result;
+        }
+        setProviders((prev) => {
+          if (!prev || !prev[provider]) return prev;
 
-        return {
-          ...prev,
-          [provider]: {
-            ...prev[provider],
-            savedCredentials: prev[provider].savedCredentials.filter(
-              (cred) => cred.id !== id,
-            ),
-          },
-        };
-      });
-      return result;
+          return {
+            ...prev,
+            [provider]: {
+              ...prev[provider],
+              savedCredentials: prev[provider].savedCredentials.filter(
+                (cred) => cred.id !== id,
+              ),
+            },
+          };
+        });
+        return result;
+      } catch (error) {
+        onFailToast("delete credentials")(error);
+        throw error;
+      }
     },
-    [api],
+    [api, onFailToast],
   );
 
   useEffect(() => {
@@ -236,45 +263,48 @@ export default function CredentialsProvider({
       return;
     }
 
-    api.listCredentials().then((response) => {
-      const credentialsByProvider = response.reduce(
-        (acc, cred) => {
-          if (!acc[cred.provider]) {
-            acc[cred.provider] = [];
-          }
-          acc[cred.provider].push(cred);
-          return acc;
-        },
-        {} as Record<CredentialsProviderName, CredentialsMetaResponse[]>,
-      );
+    api
+      .listCredentials()
+      .then((response) => {
+        const credentialsByProvider = response.reduce(
+          (acc, cred) => {
+            if (!acc[cred.provider]) {
+              acc[cred.provider] = [];
+            }
+            acc[cred.provider].push(cred);
+            return acc;
+          },
+          {} as Record<CredentialsProviderName, CredentialsMetaResponse[]>,
+        );
 
-      setProviders((prev) => ({
-        ...prev,
-        ...Object.fromEntries(
-          CREDENTIALS_PROVIDER_NAMES.map((provider) => [
-            provider,
-            {
+        setProviders((prev) => ({
+          ...prev,
+          ...Object.fromEntries(
+            CREDENTIALS_PROVIDER_NAMES.map((provider) => [
               provider,
-              providerName: providerDisplayNames[provider],
-              savedCredentials: credentialsByProvider[provider] ?? [],
-              oAuthCallback: (code: string, state_token: string) =>
-                oAuthCallback(provider, code, state_token),
-              createAPIKeyCredentials: (
-                credentials: APIKeyCredentialsCreatable,
-              ) => createAPIKeyCredentials(provider, credentials),
-              createUserPasswordCredentials: (
-                credentials: UserPasswordCredentialsCreatable,
-              ) => createUserPasswordCredentials(provider, credentials),
-              createHostScopedCredentials: (
-                credentials: HostScopedCredentialsCreatable,
-              ) => createHostScopedCredentials(provider, credentials),
-              deleteCredentials: (id: string, force: boolean = false) =>
-                deleteCredentials(provider, id, force),
-            } satisfies CredentialsProviderData,
-          ]),
-        ),
-      }));
-    });
+              {
+                provider,
+                providerName: providerDisplayNames[provider],
+                savedCredentials: credentialsByProvider[provider] ?? [],
+                oAuthCallback: (code: string, state_token: string) =>
+                  oAuthCallback(provider, code, state_token),
+                createAPIKeyCredentials: (
+                  credentials: APIKeyCredentialsCreatable,
+                ) => createAPIKeyCredentials(provider, credentials),
+                createUserPasswordCredentials: (
+                  credentials: UserPasswordCredentialsCreatable,
+                ) => createUserPasswordCredentials(provider, credentials),
+                createHostScopedCredentials: (
+                  credentials: HostScopedCredentialsCreatable,
+                ) => createHostScopedCredentials(provider, credentials),
+                deleteCredentials: (id: string, force: boolean = false) =>
+                  deleteCredentials(provider, id, force),
+              } satisfies CredentialsProviderData,
+            ]),
+          ),
+        }));
+      })
+      .catch(onFailToast("load credentials"));
   }, [
     api,
     isLoggedIn,
